+++ src/main/java/gift/controller/api/CategoryController.java
@@ -1,0 +1,25 @@
+package gift.controller.api;
+
+import gift.dto.CategoryDTO;
+import gift.service.CategoryService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/api/categories")
+public class CategoryController {
+
+    private final CategoryService categoryService;
+
+    @Autowired
+    public CategoryController(CategoryService categoryService) {
+        this.categoryService = categoryService;
+    }
+
+    @GetMapping
+    public List<CategoryDTO> getCategories() {
+        return categoryService.getAllCategories();
+    }
+}
